refactor: 설계 변경에 따른 불필요 인터페이스 제거 및 시그니처 변경
리팩토링 내용
- Rank class removed
