--- src/main/java/racingcar/model/record/Rank.java
@@ -1,4 +1,0 @@
-package racingcar.model.record;
-
-public class Rank {
-}
