+++ src/main/java/racingcar/model/record/Records.java
@@ -2,6 +2,7 @@ package racingcar.model.record;
 
 import java.util.ArrayList;
 import java.util.List;
+import racingcar.error.Error;
 
 public class Records {
     private int current_idx = 0;
@@ -16,6 +17,9 @@ public class Records {
     }
 
     public Record next() {
+        if (!hasNext()) {
+            throw new IllegalStateException(Error.HAS_NOT_NEXT_RECORD.message());
+        }
         return records.get(current_idx++);
     }
 }
